+++ apps/frontend/src/app/app.module.ts
@@ -3,12 +3,21 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './components/app/app.component';
 import { MenuComponent } from './components/menu/menu.component';
-import { ScenariiRegistryService } from './modules/shared/public-api';
+import {
+  ReportsRegistryService,
+  ScenariiRegistryService,
+} from './modules/shared/public-api';
 import { SharedModule } from './modules/shared/shared.module';
 
 const factories = {
-  loadData: (registry: ScenariiRegistryService): (() => boolean) => {
-    const f = () => registry.load();
+  loadData: (
+    scenarii: ScenariiRegistryService,
+    reports: ReportsRegistryService
+  ): (() => void) => {
+    const f = () => {
+      scenarii.load();
+      reports.load();
+    };
     return f;
   },
 };
@@ -19,7 +28,7 @@ const factories = {
     {
       provide: APP_INITIALIZER,
       multi: true,
-      deps: [ScenariiRegistryService],
+      deps: [ScenariiRegistryService, ReportsRegistryService],
       useFactory: factories.loadData,
     },
   ],