+++ apps/frontend/src/app/modules/shared/models/scenario.ts
@@ -1,8 +1,8 @@
 export interface IScenario {
   id: string;
   message: string;
-  dependency?: string;
-  comment?: string;
-  category?: string;
-  subcategory?: string;
+  dependency: string | null;
+  comment: string | null;
+  category: string | null;
+  subcategory: string | null;
 }