+++ apps/frontend/src/app/modules/shared/public-api.ts
@@ -1,5 +1,17 @@
+/**
+ * Models
+ */
 export { IScenario } from './models/scenario';
+export { IReport } from './models/report';
+
+/**
+ * Services
+ */
 export { ScenariiRegistryService } from './services/scenarii-registry/scenarii-registry.service';
 export { ReportsRegistryService } from './services/reports-registry/reports-registry.service';
 export { DataManagerService } from './services/data-manager/data-manager.service';
+
+/**
+ * Module
+ */
 export { SharedModule } from './shared.module';